backend/requirements.txt
fastapi[standard]
uvicorn
python-multipart
openai
python-dotenv
google-genai
elevenlabs
moviepy 